move file StudentsByCourseDTO and QuizCalculationResult
