+++ backend/lms/src/main/java/com/el/course/application/dto/teacher/StudentsByCourseDTO.java
@@ -1,4 +1,4 @@
-package com.el.course.domain;
+package com.el.course.application.dto.teacher;
 
 public record StudentsByCourseDTO(
         Long id,
